requirements.txt
# LLM APIs
anthropic>=0.34.0
openai>=1.40.0
google-generativeai>=0.7.0
mistralai>=1.0.0
together>=1.2.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
rich>=13.7.0
pydantic>=2.5.0

# Web search (for Phase 1 & 2)
googlesearch-python>=1.2.4
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.27.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
